Extract navbar, conditional logo, lots of style [#129]
